Support larger real values
Match distributary commit 0ca6f952ac9f131b796e94d3a8e40160211c4eb7

+++ src/convert.rs
@@ -34,7 +34,7 @@ impl ToDataType for Literal {
             Literal::FixedPoint(Real {
                 integral,
                 fractional,
-            }) => DataType::Real(integral, fractional as i32),
+            }) => DataType::Real(integral as i64, fractional as i32),
             Literal::CurrentDate => {
                 DataType::Timestamp(chrono::Local::today().and_hms(0, 0, 0).naive_local())
             }
